HereDocCtrlC

+++ src/redirect/redirect.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   redirect.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: yoav <yoav@student.42.fr>                  +#+  +:+       +#+        */
+/*   By: yrabby <yrabby@student.42.fr>              +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/26 12:46:29 by yoav              #+#    #+#             */
-/*   Updated: 2022/11/21 18:06:46 by al7aro           ###   ########.fr       */
+/*   Updated: 2022/12/11 15:19:22 by yrabby           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -29,7 +29,7 @@ void	redirect_init(t_redirect *r, char *path, t_redirect_type t)
 
 void	redirect_destroy(t_redirect **r)
 {
-	if (HEREDOC == (*r)->type)
+	if (HEREDOC == (*r)->type && (*r)->should_del_tmp)
 	{
 		unlink((*r)->path);
 		free((*r)->path);
